Statistics::Doorkeeper -> Statistics::Tasks::Doorkeeper
