+++ lib/statistics/tasks.rb
@@ -4,3 +4,4 @@ module Statistics
 end
 
 require_relative 'tasks/connpass'
+require_relative 'tasks/doorkeeper'
